feat: add health management and UI elements for player lives and score
- Introduced HealthIntValue class for managing player health with callbacks for damage and healing.
- Added Hearts and HorizontalPositiveInteger classes for displaying player lives and score in the UI.
- Updated GameSystem to integrate health management and UI elements during game initialization.

+++ src/main/java/com/github/codestorm/bounceverse/typing/interfaces/ScoreLike.java
@@ -13,14 +13,14 @@ import java.time.Instant;
  * @param <T> Loại Điểm số
  */
 public interface ScoreLike<T extends ScoreLike<T>> extends Serializable, Comparable<T> {
-    long score();
+    int score();
 
     Instant timestamp();
 
     @Override
     default int compareTo(T o) {
         if (score() != o.score()) {
-            return Long.compare(score(), o.score());
+            return Integer.compare(score(), o.score());
         }
         return -timestamp().compareTo(o.timestamp());
     }
